requirements.txt
asgiref==3.5.0
dj-database-url==0.5.0
Django==4.0.3
django-annoying==0.10.6
django-cors-headers==3.11.0
django-extensions==3.1.5
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
Pillow==9.1.0
psycopg2-binary==2.9.3
PyJWT==2.3.0
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
yapf==0.32.0
